added controllers tests

+++ CurrencyConverter.API/DTO/CurrencyResponse.cs
@@ -5,7 +5,7 @@ namespace CurrencyConverter.API.DTO
     public class CurrencyResponse
     {
         public string name { get; set; }
-        public float rate { get; set; }
+        public decimal rate { get; set; }
         public string @base { get; set; }
         public DateTime lastUpdate { get; set; } = DateTime.Now;
     }
